requirements.txt
accelerate==0.24.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
appnope==0.1.3
argilla==1.12.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
biopython==1.81
black==23.3.0
bleach==6.0.0
boto3==1.26.164
botocore==1.29.164
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.4
cloudpickle==2.2.1
comm==0.1.3
commonmark==0.9.1
contextlib2==21.6.0
contourpy==1.2.0
cryptography==41.0.1
cycler==0.12.1
Cython==0.29.36
dataclasses-json==0.5.9
datasets==2.15.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.6
distlib==0.3.6
et-xmlfile==1.1.0
evaluate==0.4.1
executing==1.2.0
fair-esm==2.0.0
faiss-cpu==1.7.4
fastapi==0.99.1
fastjsonschema==2.17.1
filelock==3.12.0
filetype==1.2.0
fonttools==4.45.1
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.5.0
google-pasta==0.2.0
greenlet==3.0.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.19.4
idna==3.4
importlib-metadata==4.13.0
install==1.3.5
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
isort==5.12.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.5
langchain==0.0.227
langchainplus-sdk==0.0.20
libmagic==1.0
llvmlite==0.41.1
lxml==4.9.3
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mistune==2.0.5
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.2.1
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numba==0.58.1
numexpr==2.8.4
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
olefile==0.46
openai==0.27.7
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
overrides==7.3.1
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathos==0.3.0
pathspec==0.11.1
pdf2image==1.16.3
pdfminer.six==20221105
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.2.0
pox==0.3.2
ppft==1.7.6.6
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==3.20.3
protobuf3-to-dict==0.1.5
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pyarrow-hotfix==0.5
pycparser==2.21
pydantic==1.10.11
Pygments==2.14.0
pynndescent==0.5.11
pypandoc==1.11
pyparsing==3.1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-json-logger==2.0.7
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.3
QtPy==2.3.1
regex==2023.5.5
requests==2.31.0
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.0.1
s3transfer==0.6.1
safetensors==0.4.0
sagemaker==2.169.0
schema==0.7.5
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.13.0
Send2Trash==1.8.2
seqeval==1.2.2
six==1.16.0
smdebug-rulesconfig==1.0.1
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.18
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tblib==1.7.0
tenacity==8.2.2
terminado==0.17.1
threadpoolctl==3.1.0
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.15.0
torch==2.1.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.36.0
triton==2.1.0
typer==0.7.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
umap-learn==0.5.5
unstructured==0.8.0
uri-template==1.2.0
urllib3==1.26.16
uvicorn==0.22.0
virtualenv==20.23.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
websockets==11.0.3
widgetsnbextension==4.0.7
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.1.2
xxhash==3.4.1
yarl==1.9.2
zipp==3.15.0